requirements.txt
fastapi
uvicorn
httpx
python-dotenv
python-multipart
python-jose[cryptography]
